\name{categories}
\alias{categories}
\docType{data}
\title{
Harvesting details related with Zhu et al 2018 data set
}
\description{
Harvesting details for all the samples selected in Data.Rd in Zhu et al 2018
}
\usage{data("categories")}
\format{
  A data frame with 402 observations on the following variable.
  \describe{
    \item{\code{Categories}}{a character vector}
  }
}
\details{
This data set was obtained from Zhu et al 2018 Supp Data
}
\source{
https://pubmed.ncbi.nlm.nih.gov/29328914/

}
\references{
Zhu et al 2018}
\examples{
data(categories)
## maybe str(categories) ; plot(categories) ...
}
\keyword{datasets}
