\name{Data_raw}
\alias{Data_raw}
\docType{data}
\title{
Metabolome changes during tomato domestication
}
\description{
Compounds detected and quantified by LC-MS/MS in different tomato and tomato-related species. NAs values are not replaced.
}
\usage{data("Data_raw")}
\format{
  A data frame with 980 observations on the following 402 variables.
  \describe{
    \item{\code{S..lycopersicum}}{a numeric vector}
    \item{\code{S..lycopersicum.1}}{a numeric vector}
    \item{\code{S..lycopersicum.2}}{a numeric vector}
    \item{\code{S..lycopersicum.3}}{a numeric vector}
    \item{\code{S..lycopersicum.4}}{a numeric vector}
    \item{\code{S..lycopersicum.5}}{a numeric vector}
    \item{\code{S..lycopersicum.6}}{a numeric vector}
    \item{\code{S..lycopersicum.7}}{a numeric vector}
    \item{\code{S..lycopersicum.8}}{a numeric vector}
    \item{\code{S..lycopersicum.9}}{a numeric vector}
    \item{\code{S..lycopersicum.10}}{a numeric vector}
    \item{\code{S..lycopersicum.11}}{a numeric vector}
    \item{\code{S..lycopersicum.12}}{a numeric vector}
    \item{\code{S..lycopersicum.13}}{a numeric vector}
    \item{\code{S..lycopersicum.14}}{a numeric vector}
    \item{\code{S..lycopersicum.15}}{a numeric vector}
    \item{\code{S..lycopersicum.16}}{a numeric vector}
    \item{\code{S..lycopersicum.17}}{a numeric vector}
    \item{\code{S..lycopersicum.18}}{a numeric vector}
    \item{\code{S..lycopersicum.19}}{a numeric vector}
    \item{\code{S..lycopersicum.20}}{a numeric vector}
    \item{\code{S..lycopersicum.21}}{a numeric vector}
    \item{\code{S..lycopersicum.22}}{a numeric vector}
    \item{\code{S..lycopersicum.23}}{a numeric vector}
    \item{\code{S..lycopersicum.24}}{a numeric vector}
    \item{\code{S..lycopersicum.25}}{a numeric vector}
    \item{\code{S..lycopersicum.26}}{a numeric vector}
    \item{\code{S..lycopersicum.27}}{a numeric vector}
    \item{\code{S..lycopersicum.28}}{a numeric vector}
    \item{\code{S..lycopersicum.29}}{a numeric vector}
    \item{\code{S..lycopersicum.30}}{a numeric vector}
    \item{\code{S..lycopersicum.31}}{a numeric vector}
    \item{\code{S..lycopersicum.32}}{a numeric vector}
    \item{\code{S..lycopersicum.33}}{a numeric vector}
    \item{\code{S..lycopersicum.34}}{a numeric vector}
    \item{\code{S..lycopersicum.35}}{a numeric vector}
    \item{\code{S..lycopersicum.36}}{a numeric vector}
    \item{\code{S..lycopersicum.37}}{a numeric vector}
    \item{\code{S..lycopersicum.38}}{a numeric vector}
    \item{\code{S..lycopersicum.39}}{a numeric vector}
    \item{\code{S..lycopersicum.40}}{a numeric vector}
    \item{\code{S..lycopersicum.41}}{a numeric vector}
    \item{\code{S..lycopersicum.42}}{a numeric vector}
    \item{\code{S..lycopersicum.43}}{a numeric vector}
    \item{\code{S..lycopersicum.44}}{a numeric vector}
    \item{\code{S..lycopersicum.45}}{a numeric vector}
    \item{\code{S..lycopersicum.46}}{a numeric vector}
    \item{\code{S..lycopersicum.47}}{a numeric vector}
    \item{\code{S..lycopersicum.48}}{a numeric vector}
    \item{\code{S..lycopersicum.49}}{a numeric vector}
    \item{\code{S..lycopersicum.50}}{a numeric vector}
    \item{\code{S..lycopersicum.51}}{a numeric vector}
    \item{\code{S..lycopersicum.52}}{a numeric vector}
    \item{\code{S..lycopersicum.53}}{a numeric vector}
    \item{\code{S..lycopersicum.54}}{a numeric vector}
    \item{\code{S..lycopersicum.55}}{a numeric vector}
    \item{\code{S..lycopersicum.56}}{a numeric vector}
    \item{\code{S..lycopersicum.57}}{a numeric vector}
    \item{\code{S..lycopersicum.58}}{a numeric vector}
    \item{\code{S..lycopersicum.59}}{a numeric vector}
    \item{\code{S..lycopersicum.60}}{a numeric vector}
    \item{\code{S..lycopersicum.61}}{a numeric vector}
    \item{\code{S..lycopersicum.62}}{a numeric vector}
    \item{\code{S..lycopersicum.63}}{a numeric vector}
    \item{\code{S..lycopersicum.64}}{a numeric vector}
    \item{\code{S..lycopersicum.65}}{a numeric vector}
    \item{\code{S..lycopersicum.66}}{a numeric vector}
    \item{\code{S..lycopersicum.67}}{a numeric vector}
    \item{\code{S..lycopersicum.68}}{a numeric vector}
    \item{\code{S..lycopersicum.69}}{a numeric vector}
    \item{\code{S..lycopersicum.70}}{a numeric vector}
    \item{\code{S..lycopersicum.71}}{a numeric vector}
    \item{\code{S..lycopersicum.72}}{a numeric vector}
    \item{\code{S..lycopersicum.73}}{a numeric vector}
    \item{\code{S..lycopersicum.74}}{a numeric vector}
    \item{\code{S..lycopersicum.75}}{a numeric vector}
    \item{\code{S..lycopersicum.76}}{a numeric vector}
    \item{\code{S..lycopersicum.77}}{a numeric vector}
    \item{\code{S..lycopersicum.78}}{a numeric vector}
    \item{\code{S..lycopersicum.79}}{a numeric vector}
    \item{\code{S..lycopersicum.80}}{a numeric vector}
    \item{\code{S..lycopersicum.81}}{a numeric vector}
    \item{\code{S..lycopersicum.82}}{a numeric vector}
    \item{\code{S..lycopersicum.83}}{a numeric vector}
    \item{\code{S..lycopersicum.84}}{a numeric vector}
    \item{\code{S..lycopersicum.85}}{a numeric vector}
    \item{\code{S..lycopersicum.86}}{a numeric vector}
    \item{\code{S..lycopersicum.87}}{a numeric vector}
    \item{\code{S..lycopersicum.88}}{a numeric vector}
    \item{\code{S..lycopersicum.89}}{a numeric vector}
    \item{\code{S..lycopersicum.90}}{a numeric vector}
    \item{\code{S..lycopersicum.91}}{a numeric vector}
    \item{\code{S..lycopersicum.92}}{a numeric vector}
    \item{\code{S..lycopersicum.93}}{a numeric vector}
    \item{\code{S..lycopersicum.94}}{a numeric vector}
    \item{\code{S..lycopersicum.95}}{a numeric vector}
    \item{\code{S..lycopersicum.96}}{a numeric vector}
    \item{\code{S..lycopersicum.97}}{a numeric vector}
    \item{\code{S..lycopersicum.98}}{a numeric vector}
    \item{\code{S..lycopersicum.99}}{a numeric vector}
    \item{\code{S..lycopersicum.100}}{a numeric vector}
    \item{\code{S..lycopersicum.101}}{a numeric vector}
    \item{\code{S..lycopersicum.102}}{a numeric vector}
    \item{\code{S..lycopersicum.103}}{a numeric vector}
    \item{\code{S..lycopersicum.104}}{a numeric vector}
    \item{\code{S..lycopersicum.105}}{a numeric vector}
    \item{\code{S..lycopersicum.106}}{a numeric vector}
    \item{\code{S..lycopersicum.107}}{a numeric vector}
    \item{\code{S..lycopersicum.108}}{a numeric vector}
    \item{\code{S..lycopersicum.109}}{a numeric vector}
    \item{\code{S..lycopersicum.110}}{a numeric vector}
    \item{\code{S..lycopersicum.111}}{a numeric vector}
    \item{\code{S..lycopersicum.112}}{a numeric vector}
    \item{\code{S..lycopersicum.113}}{a numeric vector}
    \item{\code{S..lycopersicum.114}}{a numeric vector}
    \item{\code{S..lycopersicum.115}}{a numeric vector}
    \item{\code{S..lycopersicum.116}}{a numeric vector}
    \item{\code{S..lycopersicum.117}}{a numeric vector}
    \item{\code{S..lycopersicum.118}}{a numeric vector}
    \item{\code{S..lycopersicum.119}}{a numeric vector}
    \item{\code{S..lycopersicum.120}}{a numeric vector}
    \item{\code{S..lycopersicum.121}}{a numeric vector}
    \item{\code{S..lycopersicum.122}}{a numeric vector}
    \item{\code{S..lycopersicum.123}}{a numeric vector}
    \item{\code{S..lycopersicum.124}}{a numeric vector}
    \item{\code{S..lycopersicum.125}}{a numeric vector}
    \item{\code{S..lycopersicum.126}}{a numeric vector}
    \item{\code{S..lycopersicum.127}}{a numeric vector}
    \item{\code{S..lycopersicum.128}}{a numeric vector}
    \item{\code{S..lycopersicum.129}}{a numeric vector}
    \item{\code{S..lycopersicum.130}}{a numeric vector}
    \item{\code{S..lycopersicum.131}}{a numeric vector}
    \item{\code{S..lycopersicum.132}}{a numeric vector}
    \item{\code{S..lycopersicum.133}}{a numeric vector}
    \item{\code{S..lycopersicum.134}}{a numeric vector}
    \item{\code{S..lycopersicum.135}}{a numeric vector}
    \item{\code{S..lycopersicum.136}}{a numeric vector}
    \item{\code{S..lycopersicum.137}}{a numeric vector}
    \item{\code{S..lycopersicum.138}}{a numeric vector}
    \item{\code{S..lycopersicum.139}}{a numeric vector}
    \item{\code{S..lycopersicum.140}}{a numeric vector}
    \item{\code{S..lycopersicum.141}}{a numeric vector}
    \item{\code{S..lycopersicum.142}}{a numeric vector}
    \item{\code{S..lycopersicum.143}}{a numeric vector}
    \item{\code{S..lycopersicum.144}}{a numeric vector}
    \item{\code{S..lycopersicum.145}}{a numeric vector}
    \item{\code{S..lycopersicum.146}}{a numeric vector}
    \item{\code{S..lycopersicum.147}}{a numeric vector}
    \item{\code{S..lycopersicum.148}}{a numeric vector}
    \item{\code{S..lycopersicum.149}}{a numeric vector}
    \item{\code{S..lycopersicum.150}}{a numeric vector}
    \item{\code{S..lycopersicum.151}}{a numeric vector}
    \item{\code{S..lycopersicum.152}}{a numeric vector}
    \item{\code{S..lycopersicum.153}}{a numeric vector}
    \item{\code{S..lycopersicum.154}}{a numeric vector}
    \item{\code{S..lycopersicum.155}}{a numeric vector}
    \item{\code{S..lycopersicum.156}}{a numeric vector}
    \item{\code{S..lycopersicum.157}}{a numeric vector}
    \item{\code{S..lycopersicum.158}}{a numeric vector}
    \item{\code{S..lycopersicum.159}}{a numeric vector}
    \item{\code{S..lycopersicum.160}}{a numeric vector}
    \item{\code{S..lycopersicum.161}}{a numeric vector}
    \item{\code{S..lycopersicum.162}}{a numeric vector}
    \item{\code{S..lycopersicum.163}}{a numeric vector}
    \item{\code{S..lycopersicum.164}}{a numeric vector}
    \item{\code{S..lycopersicum.165}}{a numeric vector}
    \item{\code{S..lycopersicum.166}}{a numeric vector}
    \item{\code{S..lycopersicum.167}}{a numeric vector}
    \item{\code{S..lycopersicum.168}}{a numeric vector}
    \item{\code{S..lycopersicum.169}}{a numeric vector}
    \item{\code{S..lycopersicum.170}}{a numeric vector}
    \item{\code{S..lycopersicum.171}}{a numeric vector}
    \item{\code{S..lycopersicum.172}}{a numeric vector}
    \item{\code{S..lycopersicum.173}}{a numeric vector}
    \item{\code{S..lycopersicum.174}}{a numeric vector}
    \item{\code{S..lycopersicum.175}}{a numeric vector}
    \item{\code{S..lycopersicum.176}}{a numeric vector}
    \item{\code{S..lycopersicum.177}}{a numeric vector}
    \item{\code{S..lycopersicum.178}}{a numeric vector}
    \item{\code{S..lycopersicum.179}}{a numeric vector}
    \item{\code{S..lycopersicum.180}}{a numeric vector}
    \item{\code{S..lycopersicum.181}}{a numeric vector}
    \item{\code{S..lycopersicum.182}}{a numeric vector}
    \item{\code{S..lycopersicum.183}}{a numeric vector}
    \item{\code{S..lycopersicum.184}}{a numeric vector}
    \item{\code{S..lycopersicum.185}}{a numeric vector}
    \item{\code{S..lycopersicum.186}}{a numeric vector}
    \item{\code{S..lycopersicum.187}}{a numeric vector}
    \item{\code{S..lycopersicum.188}}{a numeric vector}
    \item{\code{S..lycopersicum.189}}{a numeric vector}
    \item{\code{S..lycopersicum.190}}{a numeric vector}
    \item{\code{S..lycopersicum.191}}{a numeric vector}
    \item{\code{S..lycopersicum.192}}{a numeric vector}
    \item{\code{S..lycopersicum.193}}{a numeric vector}
    \item{\code{S..lycopersicum.194}}{a numeric vector}
    \item{\code{S..lycopersicum.195}}{a numeric vector}
    \item{\code{S..lycopersicum.196}}{a numeric vector}
    \item{\code{S..lycopersicum.197}}{a numeric vector}
    \item{\code{S..lycopersicum.198}}{a numeric vector}
    \item{\code{S..lycopersicum.199}}{a numeric vector}
    \item{\code{S..lycopersicum.200}}{a numeric vector}
    \item{\code{S..lycopersicum.201}}{a numeric vector}
    \item{\code{S..lycopersicum.202}}{a numeric vector}
    \item{\code{S..lycopersicum.203}}{a numeric vector}
    \item{\code{S..lycopersicum.204}}{a numeric vector}
    \item{\code{S..lycopersicum.205}}{a numeric vector}
    \item{\code{S..lycopersicum.206}}{a numeric vector}
    \item{\code{S..lycopersicum.207}}{a numeric vector}
    \item{\code{S..lycopersicum.208}}{a numeric vector}
    \item{\code{S..lycopersicum.209}}{a numeric vector}
    \item{\code{S..lycopersicum.210}}{a numeric vector}
    \item{\code{S..lycopersicum.211}}{a numeric vector}
    \item{\code{S..lycopersicum.212}}{a numeric vector}
    \item{\code{S..lycopersicum.213}}{a numeric vector}
    \item{\code{S..lycopersicum.214}}{a numeric vector}
    \item{\code{S..lycopersicum.215}}{a numeric vector}
    \item{\code{S..lycopersicum.216}}{a numeric vector}
    \item{\code{S..lycopersicum.217}}{a numeric vector}
    \item{\code{S..lycopersicum.218}}{a numeric vector}
    \item{\code{S..lycopersicum.219}}{a numeric vector}
    \item{\code{S..lycopersicum.220}}{a numeric vector}
    \item{\code{S..lycopersicum.221}}{a numeric vector}
    \item{\code{S..lycopersicum.222}}{a numeric vector}
    \item{\code{S..lycopersicum.223}}{a numeric vector}
    \item{\code{S..lycopersicum.224}}{a numeric vector}
    \item{\code{S..lycopersicum.225}}{a numeric vector}
    \item{\code{S..lycopersicum.226}}{a numeric vector}
    \item{\code{S..lycopersicum.227}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.1}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.2}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.3}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.4}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.5}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.6}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.7}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.8}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.9}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.10}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.11}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.12}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.13}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.14}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.15}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.16}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.17}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.18}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.19}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.20}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.21}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.22}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.23}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.24}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.25}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.26}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.27}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.28}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.29}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.30}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.31}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.32}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.33}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.34}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.35}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.36}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.37}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.38}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.39}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.40}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.41}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.42}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.43}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.44}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.45}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.46}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.47}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.48}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.49}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.50}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.51}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.52}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.53}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.54}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.55}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.56}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.57}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.58}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.59}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.60}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.61}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.62}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.63}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.64}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.65}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.66}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.67}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.68}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.69}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.70}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.71}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.72}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.73}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.74}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.75}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.76}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.77}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.78}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.79}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.80}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.81}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.82}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.83}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.84}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.85}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.86}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.87}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.88}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.89}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.90}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.91}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.92}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.93}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.94}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.95}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.96}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.97}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.98}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.99}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.100}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.101}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.102}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.103}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.104}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.105}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.106}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.107}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.108}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.109}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.110}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.111}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.112}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.113}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.114}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.115}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.116}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.117}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.118}}{a numeric vector}
    \item{\code{S..lycopersicum.var.cerasiforme.119}}{a numeric vector}
    \item{\code{S..pimpinellifolium}}{a numeric vector}
    \item{\code{S..pimpinellifolium.1}}{a numeric vector}
    \item{\code{S..pimpinellifolium.2}}{a numeric vector}
    \item{\code{S..pimpinellifolium.3}}{a numeric vector}
    \item{\code{S..pimpinellifolium.4}}{a numeric vector}
    \item{\code{S..pimpinellifolium.5}}{a numeric vector}
    \item{\code{S..pimpinellifolium.6}}{a numeric vector}
    \item{\code{S..pimpinellifolium.7}}{a numeric vector}
    \item{\code{S..pimpinellifolium.8}}{a numeric vector}
    \item{\code{S..pimpinellifolium.9}}{a numeric vector}
    \item{\code{S..pimpinellifolium.10}}{a numeric vector}
    \item{\code{S..pimpinellifolium.11}}{a numeric vector}
    \item{\code{S..pimpinellifolium.12}}{a numeric vector}
    \item{\code{S..pimpinellifolium.13}}{a numeric vector}
    \item{\code{S..pimpinellifolium.14}}{a numeric vector}
    \item{\code{S..pimpinellifolium.15}}{a numeric vector}
    \item{\code{S..pimpinellifolium.16}}{a numeric vector}
    \item{\code{S..pimpinellifolium.17}}{a numeric vector}
    \item{\code{S..pimpinellifolium.18}}{a numeric vector}
    \item{\code{S..pimpinellifolium.19}}{a numeric vector}
    \item{\code{S..pimpinellifolium.20}}{a numeric vector}
    \item{\code{S..pimpinellifolium.21}}{a numeric vector}
    \item{\code{S..pimpinellifolium.22}}{a numeric vector}
    \item{\code{S..pimpinellifolium.23}}{a numeric vector}
    \item{\code{S..pimpinellifolium.24}}{a numeric vector}
    \item{\code{S..pimpinellifolium.25}}{a numeric vector}
    \item{\code{S..pimpinellifolium.26}}{a numeric vector}
    \item{\code{S..pimpinellifolium.27}}{a numeric vector}
    \item{\code{S..pimpinellifolium.28}}{a numeric vector}
    \item{\code{S..pimpinellifolium.29}}{a numeric vector}
    \item{\code{S..pimpinellifolium.30}}{a numeric vector}
    \item{\code{S..pimpinellifolium.31}}{a numeric vector}
    \item{\code{S..pimpinellifolium.32}}{a numeric vector}
    \item{\code{S..pimpinellifolium.33}}{a numeric vector}
    \item{\code{S..pimpinellifolium.34}}{a numeric vector}
    \item{\code{S..pimpinellifolium.35}}{a numeric vector}
    \item{\code{S..pimpinellifolium.36}}{a numeric vector}
    \item{\code{S..pimpinellifolium.37}}{a numeric vector}
    \item{\code{S..pimpinellifolium.38}}{a numeric vector}
    \item{\code{S..pimpinellifolium.39}}{a numeric vector}
    \item{\code{S..pimpinellifolium.40}}{a numeric vector}
    \item{\code{S..pimpinellifolium.41}}{a numeric vector}
    \item{\code{S..pimpinellifolium.42}}{a numeric vector}
    \item{\code{S..pimpinellifolium.43}}{a numeric vector}
    \item{\code{S..pimpinellifolium.44}}{a numeric vector}
    \item{\code{S..pimpinellifolium.45}}{a numeric vector}
    \item{\code{S..pimpinellifolium.46}}{a numeric vector}
    \item{\code{S..pimpinellifolium.47}}{a numeric vector}
    \item{\code{S..pimpinellifolium.48}}{a numeric vector}
    \item{\code{S..pimpinellifolium.49}}{a numeric vector}
    \item{\code{S..pimpinellifolium.50}}{a numeric vector}
    \item{\code{S..pimpinellifolium.51}}{a numeric vector}
    \item{\code{S..pimpinellifolium.52}}{a numeric vector}
    \item{\code{S..pimpinellifolium.53}}{a numeric vector}
  }
}
\details{
The original data set was obtained from Zhu et al 2018 Supp Data
}
\source{
https://pubmed.ncbi.nlm.nih.gov/29328914/
}
\references{ Zhu et al 2018
}
\examples{
data(Data)
## maybe str(Data) ; plot(Data) ...
}
\keyword{datasets}
