% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast_ind.R
\name{Dj_index}
\alias{Dj_index}
\title{Metabolome (δj) Specialization Index}
\usage{
Dj_index(Data)
}
\arguments{
\item{Data}{a data frame the levels of differente detected  metabolites/compounds -rows-, in different samples -columns-.}
}
\value{
A data frame with the δj per species;
}
\description{
Dj_index calculates METabolome SPECialization (δj) index based on Shannon entropy.
}
\details{
Metabolome specialization δj is measured for each jth sample, as the average of the metabolites' specificities.
}
\examples{
Dj_df <- Dj_index (Data)
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
