% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast_ind.R
\name{Dj_index_plot}
\alias{Dj_index_plot}
\title{Metabolome (δj) Specialization Index Plot}
\usage{
Dj_index_plot(Dj_df)
}
\arguments{
\item{Dj_df}{a data frame with the Dj values -column- for all the samples.}
}
\value{
It returns a dot plot
}
\description{
Dj_index_plot calculates METabolome SPECialization (δj) index based on Shannon entropy.
}
\details{
Metabolome (δj) Specialization Index Plot generates a graph that shows the distrubution of the Specialization index among the samples (j).
}
\examples{
Dj_plot <- Dj_index_plot (Dj_df)
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
