% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast_sum.R
\name{Dj_index_weight}
\alias{Dj_index_weight}
\title{Metabolome Specialization index Weight Values}
\usage{
Dj_index_weight(Data, Si_df)
}
\arguments{
\item{Data}{a data frame the levels of differente detected  metabolites/compounds -rows-, in different samples -columns-.}

\item{Si_df}{a data frame with the Si values -column- for all the detected compounds - rows-, calculated from Si_index()}
}
\value{
It returns a data frame with the Pij.Si values per species and compound;
}
\description{
Dj_index_weight calculates the contribution of METaboLITE SPECialization factor (Pij*Si) to the Metabolome specialization index.
}
\details{
Metabolite specialization factor (Pij*Si) is defined as product of the a metabolite specilization index (Si) and the frequency of the metabolite in a given sample (Pij).
}
\examples{
Dj_weight_df <- Dj_index_weight (Data, Si_df)
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
