% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast_sum.R
\name{Dj_index_weight_plot}
\alias{Dj_index_weight_plot}
\title{Metabolome Specialization index Weight plot}
\usage{
Dj_index_weight_plot(Dj_weight_df)
}
\arguments{
\item{Dj_weight_df}{}
}
\value{
It returns a dot plot
}
\description{
Dj_index_weight plots the contribution of metabolite specialization factor (Pij*Si) to the Metabolome Specialization (Dj) index.
}
\details{
Metabolome Specialization index Weight plot generates a graph that shows the contribution of each metabolite to the Metabolome Specialization (Dj) index.
}
\examples{
Dj_weight_plot <- Dj_index_weight_plot (Dj_weight_df)
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
