% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast_int.R
\name{Hj_fc}
\alias{Hj_fc}
\title{Metabolic diversity internal function}
\usage{
Hj_fc(x)
}
\value{
It returns a value with the Hj value per sample
}
\description{
It performs internal metabolic diversity (Hj) calculations.
}
\details{
Performs an internal calculation necessary to obtained the metabolic diversity per sample (Hj).
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
