% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast_int.R
\name{HRj_fc}
\alias{HRj_fc}
\title{Metabolic diversity divergence internal function}
\usage{
HRj_fc(x)
}
\value{
It returns a value with the HRj value per sample
}
\description{
It performs internal metabolic diversity divergence (HRj) calculations.
}
\details{
Performs an internal calculation necessary to obtained the metabolic diversity divergence (HRj).
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
