% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast.R
\name{HRj_function}
\alias{HRj_function}
\title{Metabolic diversity divergence internal function}
\usage{
HRj_function(x)
}
\value{
It returns a dataframe
}
\description{
It performs internal metabolic diversity divergence (HRj) calculations.
}
\details{
The metabolic diversity divergence internal function performs an internal calculation necessary to obtained the metabolic diversity divergence (HRj).
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
