% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast_ind.R
\name{MetDiv}
\alias{MetDiv}
\title{Metabolic Diversity Index}
\usage{
MetDiv(Data)
}
\arguments{
\item{Data}{a data frame the levels of differente detected  metabolites/compounds -rows-, in different samples -columns-.}
}
\value{
A data frame with the Hj value and number of peaks (compounds) per species
}
\description{
MetDiv calculates METabolic Diversity (Hj) index based on Shannon entropy.
}
\details{
The metabolic profile diversity is defined as the Shannon entropy using metabolite frequency distribution in a sample (j) -Hj index-. Hj can take any value between zero when only one metabolite is detected up to log2(m), where all m metabolites are detected and accumulates at the same frequency: 1/m.
}
\examples{
Hj <- MetDiv (Data)
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
