% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast_pipe.R
\name{MetDiv_pipe}
\alias{MetDiv_pipe}
\title{Metabolic Diversity pipe}
\usage{
MetDiv_pipe(Data, ...)
}
\arguments{
\item{Data}{a data frame the levels of differente detected  metabolites/compounds -rows-, in different samples -columns-.}

\item{...}{any other argument}
}
\value{
It returns a list with 5 objects:
1. A data frame with the Hj value and number of peaks (compounds) per species;
2. A boxplot despicting the variation of Hj per species;
3. A dot plot despicting the dependency between the number of peaks and Hj
4. A dot plot despicting the dependency beween the number of peaks and Species;
5. A grid with all the plots
}
\description{
MetDiv_pipe calculates and summarises all the calculation and plots related with Metabolic Diversity.
}
\details{
MetDiv_pipe pipes the functions MetDiv(), MetDiv_plot(), and MetDiv_Supp_plot(). It takes as an input the original data set.
All the outputs can be easily unlisted, and extracted, and plotted.
}
\examples{
MetDiv_pipe <- MetDiv_pipe (Data)
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
