% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast_ind.R
\name{MetDiv_plot}
\alias{MetDiv_plot}
\title{Metabolic Diversity Plot}
\usage{
MetDiv_plot(MetDiv_df)
}
\arguments{
\item{MetDiv_df}{a data frame from MetDiv()}
}
\value{
It returns overlapped box and dot plots
}
\description{
MetDiv_plot generates a plot that shows the distribution of Hj indexes in the data set.
}
\details{
MetDiv_plot takes as an input the data frame generated by MetDiv() plotting the distribution of Hj indexes calculated by group of samples (i.e. species).
}
\examples{
Hj_plot <- MetDiv _plot(MetDiv_df)
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
