% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast_ind.R
\name{MetDiv_Supp_plot}
\alias{MetDiv_Supp_plot}
\title{Metabolic Diversity Supporting Plots}
\usage{
MetDiv_Supp_plot(MetDiv_df, y_element)
}
\arguments{
\item{MetDiv_df}{a data frame from MetDiv()}

\item{y_element}{column names from MetDiv_df: "Sample" or "Hj"}
}
\value{
It returns dot plots
}
\description{
MetDiv_Supp_plot can generates 2 plots that shows the depedency of the detect compounds with Samples or Hj.
}
\details{
MetDiv_plot takes as an input the data frame generated by MetDiv() plotting the relationship between the number of detected compounds and the samples (i.e. species), or Hj index .
}
\examples{
MetDiv_Supp_plot <- MetDiv _Supp_plot(MetDiv_df = Hj, y_element = Sample)
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
