% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast.R
\name{MetliteSpec}
\alias{MetliteSpec}
\title{Metabolite Specialization Analysis}
\usage{
MetliteSpec(Data)
}
\value{
It returns a list with 4 objects:
1. A data frame with the Pij.Si values per species and compound;
2. A data frame with the highest Pij.Si compound value per species;
3. A point plot despicting the Pij.Si value of each species
4. A point plot despicting the highest Pij.Si value per species
}
\description{
MetliteSpec calculates the contribution of METaboLITE SPECialization factor (Pij.Si) to the Metabolome specialization index.
}
\details{
Metabolite specialization factor (Pij.Si) is defined as product of the a metabolite specilization index (Si) and the frequency of the metabolite in a given sample (Pij).
}
\examples{
MetliteSpec <- MetliteSpec (Data)
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
