% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast_pipe.R
\name{MetliteSpec_pipe}
\alias{MetliteSpec_pipe}
\title{Metabolite Specialization Analysis Pipe}
\usage{
MetliteSpec_pipe(Data, ...)
}
\arguments{
\item{Data}{a data frame the levels of differente detected  metabolites/compounds -rows-, in different samples -columns-.}

\item{...}{any other argument}
}
\value{
It returns a list with 4 objects:
1. A data frame with the Pij*Si values per species and compound;
2. A data frame with the highest Pij*Si compound value per species;
3. A dot plot depicting the Pij*Si value of each species
4. A dot plot depicting the highest Pij*Si value per species
}
\description{
MetliteSpec_pipe calculates the contribution of Metabolite Specialization factor (Pij*Si) to the Metabolome Specialization (δj) index.
}
\details{
Metabolite specialization factor (Pij*Si) is defined as product of the a metabolite specilization index (Si) and the frequency of the metabolite in a given sample (Pij).
All the outputs can be easily unlisted, and extracted, and plotted.
}
\examples{
MetliteSpec_pipe <- MetliteSpec_pipe (Data)
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
