% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast.R
\name{MetPar}
\alias{MetPar}
\title{Metabolic Plasticity Parameters Summary}
\usage{
MetPar(Data)
}
\value{
It returns a data frame
}
\description{
MetPar integrates Hj, and δj indexes calculations.
}
\details{
Metabolic Plasticity Parameters Summary extracts and summarizes Hj and δj information from MetDiv and MetSpec calculations per species
}
\examples{
Hj <- MetDiv (Data)
Dj <- MetDiv (Data)
MetPar <- MetPar (Data)
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
