% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast_sum.R
\name{MetPar_df}
\alias{MetPar_df}
\title{Metabolic Plasticity Parameters}
\usage{
MetPar_df(Stats_df, Dj_df)
}
\arguments{
\item{Stats_df}{a data frame with statistical values for each samples, calculated from MetStats()}

\item{Dj_df}{a data frame with Metabolome specialization index δj calculated from Dj_index()}
}
\value{
It returns a data frame
}
\description{
MetPar integrates Hj, δj, numb_peaks, HRj, and Divj parameters
}
\details{
Metabolic Plasticity Parameters extracts and summarizes Hj, δj and statistical information.
}
\examples{
MetPar <- MetPar_df (Stat, Dj_df)
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
