% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast_pipe.R
\name{MetPar_pipe}
\alias{MetPar_pipe}
\title{Metabolic Plasticity Parameters Pipe}
\usage{
MetPar_pipe(Stats_df, Dj_df)
}
\arguments{
\item{Stats_df}{a data frame with statistical values for each samples, calculated from MetStats()}

\item{Dj_df}{a data frame with Metabolome specialization index δj calculated from Dj_index()}
}
\value{
It return a list with 3 objects:
1. A data frame with all the Metabolic Parameters per sample;
2. A dot plot depicting the relationship between the Hj and δj values of each sample
3. A dot plot depicting the relationship between Divj and δj values of each sample
}
\description{
MetPar_pipe integrates Hj, δj, numb_peaks, HRj, and Divj parameters, as well as graphically evaluate the relationship between some of them.
}
\details{
Metabolic Plasticity Parameters Pipe extracts and summarizes Hj, δj and statistical information. It generates dot plots to evaluate the relatioship between the Metabolome Specialization (δj) index and other metabolic parameters such as: Divj, and Hj.
}
\examples{
MetPar <- MetPar_pipe (Stats_df, Dj_df)
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
