% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast_sum.R
\name{MetPar_plot}
\alias{MetPar_plot}
\title{Metabolic Plasticity Parameters Plot}
\usage{
MetPar_plot(MetPar_df, x_element, ref_line = FALSE, ...)
}
\arguments{
\item{MetPar_df}{a data frame with all the metabolic plasticity paramenters obtained by MetPar_df ()}

\item{x_element}{column names from MetPar_df: "Divj", "Sample", or "Hj"}

\item{ref_line}{a logical argument. If TRUE, draws a x=y line. DEFAULT= FALSE}

\item{...}{any other argument}
}
\value{
It returns a dot plot
}
\description{
MetPar allows to graphically evaluate the relationship between Hj, δj, numb_peaks, HRj, and Divj parameters
}
\details{
Metabolic Plasticity Parameters Plot allows to generate plots to evaluate the relatioship between the Metabolome Specialization (δj) index and other metabolic parameters such as: Divj, Sample, Hj.
}
\examples{
MetPar_plot <- MetPar_plot (MetPar_df, Divj, ref_line = "TRUE")
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
