% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast.R
\name{MetSpec}
\alias{MetSpec}
\title{Metabolite (Si) and Metabolome (δj) Specialization Index}
\usage{
MetSpec(Data)
}
\value{
It returns a list with 4 objects:
1. A data frame with the Si per compound;
2. A data frame with the δj per Species;
3. A point plot despicting the Si value of each compound
4. A point plot despicting the δj per species
}
\description{
MetSpec calculates METabolite SPECialization (Si) and METabolome SPECialization (δj) indexes based on Shannon entropy.
}
\details{
Metabolic specificity (Si) is defined as the specificity of a particular MS/MS metabolite (i) among a set of samples (j). Metabolome specialization δj is measured for each jth sample, as the average of the MS/MS specificities.
}
\examples{
Dj <- MetSpec (Data)
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
