% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast_pipe.R
\name{MetSpec_pipe}
\alias{MetSpec_pipe}
\title{Metabolite (Si) and Metabolome (δj) Specialization Index Pipeline}
\usage{
MetSpec_pipe(Data, ...)
}
\arguments{
\item{Data}{a data frame the levels of differente detected  metabolites/compounds -rows-, in different samples -columns-.}

\item{...}{any other argument}
}
\value{
It returns a list with 4 objects:
1. A data frame with the Si per compound;
2. A data frame with the δj per Species;
3. A dot plot depicting the Si value of each compound
4. A dot plot depicting the δj per species
}
\description{
MetSpec_pipe calculates METabolite SPECialization (Si) and METabolome SPECialization (δj) indexes based on Shannon entropy.
}
\details{
Metabolic specificity (Si) is defined as the specificity of a particular metabolite (i) among a set of samples (j).
Metabolome specialization δj is measured for each jth sample, as the average of the metabolites specificity.
All the outputs can be easily unlisted, and extracted, and plotted.
}
\examples{
Dj <- MetSpec_pipe (Data)
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
