% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast_sum.R
\name{MetStats}
\alias{MetStats}
\title{Metabolic Plasticity Statistics}
\usage{
MetStats(Data, MetDiv_df)
}
\arguments{
\item{Data}{a data frame the levels of differente detected  metabolites/compounds -rows-, in different samples -columns-.}

\item{MetDiv_df}{a data frame from MetDiv()}
}
\value{
It returns a data frame with the all the species-related statistical parameters: Samples, Hj, numb_peaks, HRj, and Divj.
}
\description{
MetStats calculates sample (HRj) and Kullback–Leibler divergence (Divj)
}
\details{
Metabolic Plasticity Statistics calculates HRj and Divj considering Hj calculations per species
}
\examples{
Stats <- MetStats (Data, MetDiv_df)
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
