% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast_int.R
\name{my_filter_fc}
\alias{my_filter_fc}
\title{Metabolite contribution to metabolome specialization index (δj) function}
\usage{
my_filter_fc(Dj_index_weight)
}
\value{
It returns a table
}
\description{
It selects the metabolites that contribute the most to the metabolome specialization index (δj) - i.e. highest Pij*Si
}
\details{
Metabolite contribution to metabolome specialization index (δj) function select the metabolite with the highest Pij*Si value per samples (i.e. species).
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
