% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast.R
\name{my_filter_function}
\alias{my_filter_function}
\title{Metabolite contribution to metabolome specialization index (δj) function}
\usage{
my_filter_function(table)
}
\value{
It returns a table
}
\description{
It selects the metabolites that contribute the most to the metabolome specialization index (δj).
}
\details{
Metabolite contribution to metabolome specialization index (δj) function select the metabolite with the highest Pij*Si value per species.
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
