% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast_int.R
\name{numb_peaks_fc}
\alias{numb_peaks_fc}
\title{Number of detected compounds function}
\usage{
numb_peaks_fc(Data_raw)
}
\arguments{
\item{Data_raw}{a data frame the levels of different detected  metabolites/compounds -rows-, in different samples -columns-, without replacing the NAs.}
}
\value{
It returns a vector with the number of peaks - e.g. detected compounds - per species.
}
\description{
It counts the number of detected compounds per sample
}
\details{
The function identified and excludes the NAs on each col, to count the total number of compounds truely identified in each sample
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
