% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast_int.R
\name{Pij_by_Pi_fc}
\alias{Pij_by_Pi_fc}
\title{Metabolite expression identity internal function}
\usage{
Pij_by_Pi_fc(Data)
}
\arguments{
\item{Data}{a data frame the levels of differente detected  metabolites/compounds -rows-, in different samples -columns-.}
}
\value{
It returns a data frame with Pij/Pi values
}
\description{
The expression identity of a given metabolite profile regarding frequencies among the set of samples (j).
}
\details{
The metabolite expression identity is calculated dividing the metabolite frequency per sample by the frequency of the given metabolite in the whole dataset
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
