% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast_int.R
\name{Pij_fc}
\alias{Pij_fc}
\title{Metabolite frequency per sample - Internal function}
\usage{
Pij_fc(Data)
}
\arguments{
\item{Data}{a data frame the levels of different detected  metabolites/compounds -rows-, in different samples -columns-.}
}
\value{
It returns a dataframe with metabolite frequencies
}
\description{
Internal function to calculate the frequency of metabolite i in sample j (Pij)
}
\details{
Calculates the frequency dividing the intensity of a given metabolite (i) by the sum of intensities of all metabolites detected in a given sample (i.e. species) (j).
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
