% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast_ind.R
\name{Si_index}
\alias{Si_index}
\title{Metabolite (Si) Specialization Index}
\usage{
Si_index(Data)
}
\arguments{
\item{Data}{a data frame the levels of differente detected  metabolites/compounds -rows-, in different samples -columns-.}
}
\value{
It returns a data frame with the Si value per compound
}
\description{
Si_index calculates METabolite SPECialization (Si) index based on Shannon entropy.
}
\details{
Metabolic specificity (Si) is defined as the specificity of a particular metabolite (i) among a set of samples (j).
}
\examples{
Si_df <- Si_index (Data)
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
