% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast_ind.R
\name{Si_index_plot}
\alias{Si_index_plot}
\title{Metabolite (Si) Specialization Index Plot}
\usage{
Si_index_plot(Si_df)
}
\arguments{
\item{Si_df}{a data frame with the Si values -column- for all the detected compounds - rows-, calculated from Si_index()}
}
\value{
It returns a dot plot
}
\description{
Si_index_plot plots the METabolite SPECialization (Si) indexes based on Shannon entropy for each compound.
}
\details{
Metabolic specificity (Si) plot generates a graph that shows the distrubution of Si index among a the set of detected compounds (i).
}
\examples{
Si_plot <- Si_index_plot (Si_df)
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
