% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast.R
\name{Hj_fc}
\alias{Hj_fc}
\title{Metabolic diversity internal function}
\usage{
Hj_fc(x)
}
\value{
It returns a dataframe
}
\description{
It performs internal metabolic diversity (Hj) calculations.
}
\details{
The metabolic diversity internal function performs an internal calculation necessary to obtained the metabolic diversity per species (Hj).
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
