% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast.R
\name{MetDiv}
\alias{MetDiv}
\title{Metabolic Diversity Index (Hj)}
\usage{
MetDiv(Data)
}
\value{
It returns a list with 5 objects:
1. A data frame with the Hj value and number of peaks (compounds) per species;
2. A boxplot despicting the variation of Hj per species;
3. A point plot despicting the dependency between the number of peaks and Hj
4. A point plot despicting the dependency beween the number of peaks and Species;
5. A grid with all the plots
}
\description{
MetDiv calculates METabolic Diversity (Hj) index based on Shannon entropy.
}
\details{
The metabolic profile diversity is defined as the Shannon entropy using MS/MS metabolite frequency distribution in a sample (j) -Hj index- (Martinez et al paper). Hj can take any value between zero when only one metabolite is detected up to log2(m), where all m metabolites are detected and accumulates at the same frequency: 1/m (Martinez and Reyes-Valdes, 2008).
}
\examples{
Hj <- MetDiv (Data)
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
