% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast.R
\name{MetStats}
\alias{MetStats}
\title{Metabolic Plasticity Statistics}
\usage{
MetStats(Data)
}
\value{
It returns a data frame
}
\description{
MetStats calculates sample (HRj) and Kullback–Leibler divergence (Divj)
}
\details{
Metabolic Plasticity Statistics calculates HRj and DivJ from MetDiv calculations per species
}
\examples{
Hj <- MetDiv (Data)
MetPar <- MetPar (Data)
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
