% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast_int.R
\name{numb_peaks_fc}
\alias{numb_peaks_fc}
\title{Number of detected compounds function}
\usage{
numb_peaks_fc(Data)
}
\value{
It returns a vector with the number of peaks - e.g. detected compounds - per species.
}
\description{
It counts the number of detected compounds per sample
}
\details{
It counts the number of compound per sample that have a value higher to the mininum.
Initially, NA values are replaced by the minimum value of the whole data set divided by 1000000.
Thus, the function considers the minimum value of the corrected data set as a NA in the original data set (i.e. absence of the compound)
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
