% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast.R
\name{numb_peaks_fc}
\alias{numb_peaks_fc}
\title{Number of detected compounds function}
\usage{
numb_peaks_fc(Data)
}
\value{
It returns a vector with the number of peaks - e.g. detected compounds - per species.
}
\description{
It counts the number of detected compounds per species
}
\details{
Number of detected compounds function counts the number of compound per species that have a value higher to the mininum. (Initially, NA values are replaced by the minimum value of the whole data set divided by 1000000 )
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
