% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast.R
\name{Pij_by_Pi_fc}
\alias{Pij_by_Pi_fc}
\title{Metabolite expression identity internal function}
\usage{
Pij_by_Pi_fc(Data)
}
\value{
It returns a dataframe with metabolite expression identity
}
\description{
The expression identity of a given MS/MS regarding frequencies among considered samples.
}
\details{
The metabolite expression identity is calculated dividing the metabolite frequency per sample by the frequency of the given metabolite in the whole dataset
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
