% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast.R
\name{Pij_fc}
\alias{Pij_fc}
\title{Metabolite frequency per sample internal function}
\usage{
Pij_fc(Data)
}
\value{
It returns a dataframe with metabolite frequency
}
\description{
Internal function to calculate the frequency of metabolite i in species j (Pij)
}
\details{
The frequency is calculated dividing the intensity of a given metabolite (i) by the sum of intensities in a given species (j).
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
