% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPlast.R
\name{Si_fc}
\alias{Si_fc}
\title{Metabolite Specialization internal function}
\usage{
Si_fc(x)
}
\value{
It returns a dataframe
}
\description{
It performs internal metabolite specialization (Si) calculations.
}
\details{
The metabolite specialization internal function performs an internal calculation necessary to obtained the metabolite specialization (Si).
}
\author{
Lucio D'Andrea, PhD; Prof Aureliano Bombarely
}
